Move icons into Assets file, Create CurrentTaco file
