+++ src/CurrentTaco/CurrentTaco.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import taco1 from '../Assets/taco1.svg';
+
+
+
+const CurrentTaco = () => {
+  return (
+    <img src={taco1} className="App-logo" alt="logo" />
+  )
+}
+
+export default CurrentTaco;
